[Feat] #44 - OnboardingUnivView Coordinator 연결
